# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get notifications for current user (admin sees all)",
        "tags": [
          "Notifications"
        ]
      },
      "post": {
        "operationId": "NotificationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Admin: create a notification",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/{id}": {
      "get": {
        "operationId": "NotificationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one notification by ID",
        "tags": [
          "Notifications"
        ]
      },
      "patch": {
        "operationId": "NotificationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a notification (read status, message, etc.)",
        "tags": [
          "Notifications"
        ]
      },
      "delete": {
        "operationId": "NotificationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Admin: delete a notification",
        "tags": [
          "Notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "Parts Management API",
    "description": "Endpoints to manage parts, claims, and users",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "NotificationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Notification UUID",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "description": "UUID of the user receiving this notification",
            "format": "uuid"
          },
          "message": {
            "type": "string",
            "description": "Notification message text"
          },
          "type": {
            "type": "string",
            "description": "Category of notification",
            "enum": [
              "PART_AVAILABLE",
              "PART_REMOVED",
              "PART_CLAIMED",
              "PART_COLLECTED",
              "GENERAL"
            ],
            "example": "PART_AVAILABLE"
          },
          "isRead": {
            "type": "boolean",
            "description": "Read status",
            "default": false
          },
          "metadata": {
            "type": "object",
            "description": "Arbitrary metadata payload",
            "additionalProperties": true
          },
          "timestamp": {
            "format": "date-time",
            "type": "string",
            "description": "When this notification was created"
          }
        },
        "required": [
          "id",
          "userId",
          "message",
          "type",
          "isRead",
          "timestamp"
        ]
      },
      "CreateNotificationDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "PART_AVAILABLE",
              "PART_REMOVED",
              "PART_CLAIMED",
              "PART_COLLECTED",
              "GENERAL"
            ]
          },
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "description": "Arbitrary metadata payload"
          },
          "isRead": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "userId",
          "type",
          "message"
        ]
      },
      "UpdateNotificationDto": {
        "type": "object",
        "properties": {
          "isRead": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "type": {
            "type": "string",
            "enum": [
              "PART_AVAILABLE",
              "PART_REMOVED",
              "PART_CLAIMED",
              "PART_COLLECTED",
              "GENERAL"
            ]
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "description": "Arbitrary metadata payload"
          }
        }
      },
...
    }
  }
}